changed the page routing approach
