+++ src/data/project/stock-exchange.md
@@ -1,7 +1,6 @@
 ---
 templateKey: project-page
 title: Stock Exchange
-slug: stock-exchange
 image: /images/project-img-4.png
 description: Lorem ipsum dolor sit amet, consectetur adipiscing elit. Fusce at
   nibh placerat, fermentum risus nec, faucibus nisi. Cras cursus ullamcorper est
